requirements.txt
opencv-python
ultralytics
deep_sort_realtime
PyYAML
streamlit  # or gradio
motmetrics